Bug 1464400 - Keep track of CSS generics when resolving to actual font families and faces, and expose as a new CSSGeneric attribute on InspectorFontFace. r=jwatt
UltraBlame original commit: f8dbb1d2d07cb5645d8370b1f766b21da538d392

+++ gfx/thebes/gfxFontFamilyList.h
@@ -23,7 +23,7 @@ namespace mozilla {
 
  
 
-enum FontFamilyType : uint32_t {
+enum FontFamilyType : uint8_t {
   eFamily_none = 0,
